Added a functions set for accessing files using buffered (threaded) input and use those functions to read mp3, ogg, and flac files.
git-svn-id: svn://svn.daper.net/moc/trunk@1056 910807d9-36e0-0310-a014-e9ea483e2ba4

+++ out_buf.h
@@ -39,7 +39,7 @@ void out_buf_unpause (struct out_buf *buf);
 void out_buf_stop (struct out_buf *buf);
 void out_buf_reset (struct out_buf *buf);
 void out_buf_time_set (struct out_buf *buf, const float time);
-float out_buf_time_get (struct out_buf *buf);
+int out_buf_time_get (struct out_buf *buf);
 void out_buf_set_notify_cond (struct out_buf *buf, pthread_cond_t *cond,
 		pthread_mutex_t *mutex);
 int out_buf_get_free (struct out_buf *buf);
